refactor: use barrel export for diff module
- Change imports from diff/hasher.js to diff/index.js
- Align implementation with module design intent
- Affected files: src/crawler/index.ts, src/output/writer.ts

Refs #864
